Appearence and recipe content updates
Replaced the logo font to Babylonica.
Made some corrections to the text following my english teacher's review.
Added some comments to the script for easier code readability.
Added more space to the #nav offset when clicked, this update was necessary due to logo size changing.
Added svg images to represent tools and ingrediets needed for the recipe making in a more visual way.